Send 403 if book does not belong to user

+++ delete.php
@@ -6,7 +6,7 @@ session_start();
 
 $id = $_GET["id"];
 
-$statement = $conn->prepare("Select * from books where id = :id");
+$statement = $conn->prepare("Select * from books where id = :id limit 1");
 $statement->execute([":id" => $id]);
 
 if ($statement->rowCount() == 0) {
@@ -14,6 +14,15 @@ if ($statement->rowCount() == 0) {
   echo("HTTP 404 NOT FOUND");
   return;
 }
+
+$contact = $statement->fetch(PDO::FETCH_ASSOC);
+
+if ($contact["user_id"] !== $_SESSION["user"]["id"]) {
+  http_response_code(403);
+  echo("HTTP 403 UNAUTHORIZED");
+  return;
+}
+
 
 $conn->prepare("Delete from books where id = :id")->execute([":id" => $id]);
